requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.3.1
pytest-cov>=4.1.0
